Back out rev 2acd6833, bug 698519 - update to libjpeg 1.2, due to Dromaeo DOM regression.
UltraBlame original commit: 075fe28b4d34bd4e0ebe868f2aba134978743db8

+++ media/libjpeg/jutils.c
@@ -77,8 +77,8 @@ jdiv_round_up (long a, long b)
 }
 
 
-GLOBAL(long)
-jround_up (long a, long b)
+GLOBAL(size_t)
+jround_up (size_t a, size_t b)
 
 
 {
